Added prompt for user input and functions
The prompt is added for user to insert a value whereas the functions is for the formation of grid within a container. The div are added into the container based on value inserted by the user. The console log shows that the function is working. Create element for div will be added

[Ticket: #3 Date: 16/6/2021]